fix: navigate back before factory reset to prevent stuck screen and token race
RemoveAccountConfirmationScreen no longer calls factoryReset directly.
Instead it navigates back to the Account tab with a removeAccount param,
so the reset runs with no pushed screen on the stack — preventing React
Navigation from entering a bad state during the RootStack tree swap.

Guards useFocusEffect to skip refreshIdToken when removal is pending,
fixing an intermittent race where the token refresh races with
deleteRegistration and the server rejects the stale token (error 215).
