+++ app/src/bcsc-theme/features/account/RemoveAccountConfirmationScreen.tsx
@@ -1,11 +1,10 @@
-import { useFactoryReset } from '@/bcsc-theme/api/hooks/useFactoryReset'
-import { BCSCMainStackParams } from '@/bcsc-theme/types/navigators'
-import { Button, ButtonType, ThemedText, TOKENS, useServices, useTheme } from '@bifold/core'
+import { BCSCMainStackParams, BCSCScreens, BCSCStacks } from '@/bcsc-theme/types/navigators'
+import { Button, ButtonType, ThemedText, useTheme } from '@bifold/core'
 import { useNavigation } from '@react-navigation/native'
 import { StackNavigationProp } from '@react-navigation/stack'
-import React, { useState } from 'react'
+import React from 'react'
 import { useTranslation } from 'react-i18next'
-import { ActivityIndicator, ScrollView, StyleSheet, View } from 'react-native'
+import { ScrollView, StyleSheet, View } from 'react-native'
 import { SafeAreaView } from 'react-native-safe-area-context'
 
 type AccountNavigationProp = StackNavigationProp<BCSCMainStackParams>
@@ -13,19 +12,17 @@ type AccountNavigationProp = StackNavigationProp<BCSCMainStackParams>
 /**
  * Screen that confirms the user's intent to remove their account.
  *
- * Uses local loading state instead of the global BCSCLoadingProvider overlay,
- * because the global overlay persists above RootStack and gets stuck when
- * the navigator tree swaps during factory reset.
+ * Does not call factoryReset directly — instead navigates back to the Account
+ * tab with a param signal, so the reset runs without this screen on the stack.
+ * This avoids React Navigation getting into a bad state when RootStack swaps
+ * navigator trees during factory reset.
  *
  * @returns {*} {React.ReactElement} The RemoveAccountConfirmationScreen component.
  */
 const RemoveAccountConfirmationScreen: React.FC = () => {
-  const { Spacing, ColorPalette } = useTheme()
+  const { Spacing } = useTheme()
   const navigation = useNavigation<AccountNavigationProp>()
   const { t } = useTranslation()
-  const factoryReset = useFactoryReset()
-  const [logger] = useServices([TOKENS.UTIL_LOGGER])
-  const [isRemoving, setIsRemoving] = useState(false)
 
   const styles = StyleSheet.create({
     container: {
@@ -41,24 +38,7 @@ const RemoveAccountConfirmationScreen: React.FC = () => {
       gap: Spacing.md,
       marginTop: Spacing.lg,
     },
-    loadingContainer: {
-      flex: 1,
-      justifyContent: 'center',
-      alignItems: 'center',
-      gap: Spacing.md,
-    },
   })
-
-  if (isRemoving) {
-    return (
-      <SafeAreaView style={styles.container} edges={['bottom', 'left', 'right']}>
-        <View style={styles.loadingContainer}>
-          <ActivityIndicator size="large" color={ColorPalette.brand.primary} />
-          <ThemedText>{t('BCSC.Account.RemoveAccountLoading')}</ThemedText>
-        </View>
-      </SafeAreaView>
-    )
-  }
 
   return (
     <SafeAreaView style={styles.container} edges={['bottom', 'left', 'right']}>
@@ -71,21 +51,11 @@ const RemoveAccountConfirmationScreen: React.FC = () => {
           accessibilityLabel={t('BCSC.Account.RemoveAccount')}
           buttonType={ButtonType.Critical}
           title={t('BCSC.Account.RemoveAccount')}
-          onPress={async () => {
-            try {
-              setIsRemoving(true)
-
-              logger.info('[RemoveAccount] User confirmed account removal, proceeding with verification reset')
-
-              const result = await factoryReset()
-
-              if (!result.success) {
-                logger.error('[RemoveAccount] Failed to remove account', result.error)
-              }
-            } catch (error) {
-              logger.error('[RemoveAccount] Error during account removal', error as Error)
-              setIsRemoving(false)
-            }
+          onPress={() => {
+            navigation.navigate(BCSCStacks.Tab, {
+              screen: BCSCScreens.Account,
+              params: { removeAccount: true },
+            })
           }}
         />
         <Button
